Testing_Best_Practices.md minor styling
Mostly minor tweak to the best practice guide in the context of a non
native english speaker.

Rephrase the introduction paragraph to be lest tied to Etsy codebase
since the guide is now publicly available on Github, thank you!

Remove most uses of apostrophes solely on the basis I find it more
explicit to use the whole word to avoid potential confusion. For
exammple `you'r`, sounds like `your` which causes additional brain
cycles to understand that it is really `you are`. If you disagree, I
will be happy to restore them.

Replace a few 'it' references with the term referened, to ease reading.

Add code comments to help developers discovering what unit testing is.